# docs/swagger.yaml
definitions:
...
  cms_internal_app_menu_internal_dto.ApiPagedData-dto_Menu:
    properties:
      list:
        items:
          $ref: '#/definitions/dto.Menu'
        type: array
      pageNo:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
...
  datatype.EnumValue:
    properties:
      label:
        type: string
      value:
        type: string
    type: object
...
  dto.ApiArticleModel:
    properties:
      code:
        type: string
      data:
        $ref: '#/definitions/dto.ArticleModel'
      msg:
        type: string
    type: object
  dto.ApiArticleModelList:
    properties:
      code:
        type: string
      data:
        $ref: '#/definitions/dto.ApiPaginatedData-dto_ArticleModel'
      msg:
        type: string
    type: object
...
  dto.ApiMenu:
    properties:
      code:
        type: string
      data:
        $ref: '#/definitions/dto.Menu'
      msg:
        type: string
    type: object
  dto.ApiMenuList:
    properties:
      code:
        type: string
      data:
        $ref: '#/definitions/cms_internal_app_menu_internal_dto.ApiPagedData-dto_Menu'
      msg:
        type: string
    type: object
  dto.ApiPaginatedData-dto_ArticleModel:
    properties:
      list:
        items:
          $ref: '#/definitions/dto.ArticleModel'
        type: array
      pageNo:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
...
  dto.ArticleModel:
    properties:
      createdAt:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      updatedAt:
        type: string
    type: object
  dto.ArticleModelCreateParams:
    properties:
      description:
        type: string
      name:
        type: string
    required:
    - description
    - name
    type: object
  dto.ArticleModelListParams:
    properties:
      name:
        type: string
      pageNo:
        type: integer
      pageSize:
        type: integer
    type: object
...
  dto.ArticleModelSchemaParams:
    properties:
      description:
        maxLength: 255
        type: string
      enable:
        type: boolean
      enumOptions:
        items:
          $ref: '#/definitions/datatype.EnumValue'
        type: array
      fieldKey:
        maxLength: 255
        type: string
      fieldName:
        maxLength: 255
        type: string
      hidden:
        type: boolean
      id:
        type: string
      maxLen:
        type: string
      maxTime:
        type: string
      maxValue:
        type: integer
      minLen:
        type: string
      minTime:
        type: string
      minValue:
        type: integer
      pattern:
        type: string
      required:
        type: boolean
      sequence:
        type: integer
      type:
        type: integer
    required:
    - fieldKey
    - fieldName
    - type
    type: object
  dto.ArticleModelSchemaUpdateParams:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.ArticleModelSchemaParams'
        type: array
      id:
        example: "123456"
        type: string
    required:
    - id
    type: object
  dto.ArticleModelUpdateParams:
    properties:
      description:
        type: string
      id:
        example: "123456"
        type: string
      name:
        type: string
    required:
    - id
    type: object
...
  dto.FileInfo:
    properties:
      createdAt:
        type: string
      id:
        type: string
      name:
        type: string
      size:
        type: integer
      updatedAt:
        type: string
      url:
        type: string
    type: object
  dto.Menu:
    properties:
      affix:
        type: integer
      createdAt:
        type: string
      externalUrl:
        type: string
      icon:
        type: string
      id:
        type: string
      name:
        type: string
      parentId:
        type: integer
      remark:
        type: string
      sequence:
        type: integer
      template:
        type: string
      type:
        type: integer
      updatedAt:
        type: string
      uri:
        type: string
      visible:
        type: boolean
    type: object
  dto.MenuCreateParams:
    properties:
      affix:
        type: integer
      externalUrl:
        type: string
      icon:
        type: string
      name:
        type: string
      parentId:
        type: string
      remark:
        type: string
      sequence:
        type: string
      template:
        type: string
      type:
        type: integer
      uri:
        type: string
    required:
    - name
    - uri
    type: object
  dto.MenuListQueryParams:
    properties:
      name:
        type: string
      pageNo:
        type: integer
      pageSize:
        type: integer
      parentId:
        type: string
    type: object
  dto.MenuMoveParams:
    properties:
      id:
        type: string
      targetId:
        type: string
    required:
    - id
    - targetId
    type: object
  dto.MenuUpdateParams:
    properties:
      affix:
        type: integer
      externalUrl:
        type: string
      icon:
        type: string
      id:
        type: string
      name:
        type: string
      remark:
        type: string
      sequence:
        type: string
      template:
        type: string
      type:
        type: integer
      uri:
        type: string
    required:
    - id
    type: object
...
  erroz.Result:
    properties:
      code:
        type: string
      data: {}
      msg:
        type: string
    type: object
...
  types.ApiCreateResult:
    properties:
      code:
        type: string
      data:
        description: 数据 ID
        example: "123456"
        type: string
      msg:
        type: string
    type: object
  types.ApiEmptyResult:
    properties:
      code:
        type: string
      msg:
        type: string
    type: object
  types.ResourceID:
    properties:
      id:
        example: "123456"
        type: string
    required:
    - id
    type: object
info:
  contact: {}
  license:
    name: Apache 2.0
  title: EgoCms Api
  version: "0.1"
paths:
  /article-model/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: 请求参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ArticleModelCreateParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.ApiCreateResult'
      security:
      - ApiKeyAuth: []
      summary: 创建文章模型
      tags:
      - 文章模型
      x-apifox-folder: 文章模型
  /article-model/delete:
    post:
      consumes:
      - application/json
      parameters:
      - description: 请求参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/types.ResourceID'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.ApiEmptyResult'
      security:
      - ApiKeyAuth: []
      summary: 删除文章模型
      tags:
      - 文章模型
      x-apifox-folder: 文章模型
  /article-model/detail:
    post:
      consumes:
      - application/json
      parameters:
      - description: 请求参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/types.ResourceID'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ApiArticleModel'
      security:
      - ApiKeyAuth: []
      summary: 查看文章模型
      tags:
      - 文章模型
      x-apifox-folder: 文章模型
  /article-model/list:
    post:
      consumes:
      - application/json
      parameters:
      - description: 请求参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ArticleModelListParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ApiArticleModelList'
      security:
      - ApiKeyAuth: []
      summary: 查看文章模型列表
      tags:
      - 文章模型
      x-apifox-folder: 文章模型
  /article-model/update:
    post:
      consumes:
      - application/json
      parameters:
      - description: 请求参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ArticleModelUpdateParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.ApiEmptyResult'
      security:
      - ApiKeyAuth: []
      summary: 更新文章模型
      tags:
      - 文章模型
      x-apifox-folder: 文章模型
  /article-model/update-schema:
    post:
      consumes:
      - application/json
      parameters:
      - description: 请求参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ArticleModelSchemaUpdateParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.ApiEmptyResult'
      security:
      - ApiKeyAuth: []
      summary: 更新文章模型Schema
      tags:
      - 文章模型
      x-apifox-folder: 文章模型
...
  /file/delete:
    post:
      consumes:
      - application/json
      parameters:
      - description: 请求参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/types.ResourceID'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.ApiEmptyResult'
      security:
      - ApiKeyAuth: []
      summary: 文件删除
      tags:
      - 文件
      x-apifox-folder: 文件
  /file/download:
    get:
      consumes:
      - application/json
      parameters:
      - description: 文件ID
        in: gquery
        name: id
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            type: file
      security:
      - ApiKeyAuth: []
      summary: 文件下载
      tags:
      - 文件
      x-apifox-folder: 文件
  /file/upload:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: 请求参数
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/erroz.Result'
            - properties:
                data:
                  $ref: '#/definitions/dto.FileInfo'
              type: object
      security:
      - ApiKeyAuth: []
      summary: 文件上传
      tags:
      - 文件
      x-apifox-folder: 文件
  /menu/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: 请求参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.MenuCreateParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.ApiEmptyResult'
      security:
      - ApiKeyAuth: []
      summary: 创建菜单
      tags:
      - 菜单
      x-apifox-folder: 菜单
  /menu/delete:
    post:
      consumes:
      - application/json
      parameters:
      - description: 请求参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/types.ResourceID'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.ApiEmptyResult'
      security:
      - ApiKeyAuth: []
      summary: 删除菜单
      tags:
      - 菜单
      x-apifox-folder: 菜单
  /menu/detail:
    post:
      consumes:
      - application/json
      parameters:
      - description: 请求参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/types.ResourceID'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ApiMenu'
      security:
      - ApiKeyAuth: []
      summary: 查看菜单
      tags:
      - 菜单
      x-apifox-folder: 菜单
  /menu/list:
    post:
      consumes:
      - application/json
      parameters:
      - description: 请求参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.MenuListQueryParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ApiMenuList'
      security:
      - ApiKeyAuth: []
      summary: 菜单列表
      tags:
      - 菜单
      x-apifox-folder: 菜单
  /menu/move:
    post:
      consumes:
      - application/json
      parameters:
      - description: 请求参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.MenuMoveParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.ApiEmptyResult'
      security:
      - ApiKeyAuth: []
      summary: 移动菜单
      tags:
      - 菜单
      x-apifox-folder: 菜单
  /menu/update:
    post:
      consumes:
      - application/json
      parameters:
      - description: 请求参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.MenuUpdateParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.ApiEmptyResult'
      security:
      - ApiKeyAuth: []
      summary: 更新菜单
      tags:
      - 菜单
      x-apifox-folder: 菜单
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"